fix: determine is_url from filename

+++ fury/io.py
@@ -7,7 +7,7 @@ from fury.utils import set_input
 from urllib.request import urlretrieve
 
 
-def load_image(filename, is_url=False, as_vtktype=False, use_pillow=True):
+def load_image(filename, as_vtktype=False, use_pillow=True):
     """Load an image.
 
     Parameters
@@ -27,6 +27,8 @@ def load_image(filename, is_url=False, as_vtktype=False, use_pillow=True):
         desired image array
 
     """
+    is_url = filename.startswith('http://') or filename.startswith('https://')
+
     if is_url:
         image_name = os.path.basename(filename)
         urlretrieve(filename, image_name)
@@ -72,7 +74,8 @@ def load_image(filename, is_url=False, as_vtktype=False, use_pillow=True):
             vtk_image.GetPointData().SetScalars(uchar_array)
             image = vtk_image
 
-        os.remove(filename)
+        if is_url:
+            os.remove(filename)
         return image
 
     d_reader = {".png": vtk.vtkPNGReader,
@@ -101,7 +104,8 @@ def load_image(filename, is_url=False, as_vtktype=False, use_pillow=True):
         image = numpy_support.vtk_to_numpy(vtk_array).reshape(h, w, components)
         image = np.flipud(image)
 
-    os.remove(filename)
+    if is_url:
+        os.remove(filename)
     return reader.GetOutput() if as_vtktype else image
